+++ src/menu.rs
@@ -123,6 +123,7 @@ const VIEW_EDITOR: &[Item] = &[
     Item::leaf("menu.item.view.whitespace", "view.whitespace", ""),
     Item::leaf("menu.item.view.scrollbar", "view.scrollbar", ""),
     Item::leaf("menu.item.view.soft_wrap", "view.soft_wrap", ""),
+    Item::leaf("menu.item.view.spellcheck", "view.spellcheck", ""),
 ];
 
 const VIEW: &[Item] = &[